move config model into its own file
