+++ OmniFarm/OmniFarm.csproj
@@ -58,6 +58,7 @@
   <ItemGroup>
     <Compile Include="MyAssetLoader.cs" />
     <Compile Include="OmniFarm.cs" />
+    <Compile Include="OmniFarmConfig.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
